requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
pdfplumber
groq
python-dotenv